fix(core): re-expose strip-source-code
